SDL: Fix an invalid memory access.
The array was being accessed with `newDisplayCount - 1` while
`newDisplayCount` was 0. Presumably, `d` originally counted upwards
instead of downwards, so instead I'm checking `newDisplayCount`,
which does still count up. Alternatively, you could check if `d` is
equal to `displayCount - 1`, but I think that this looks better and
makes the intent more obvious to the reader.