Add virtual destructors to some base structs.
Freeing an object with a pointer to a base type will freak out ASAN
unless the base type has a virtual destructor. I'm not much of a C++
guru, but I think this makes sense: `delete` seems to be basically
calling a destructor pointer, and unless it's virtual then it will
not call the correct destructor.