Improve memory allocation.
I believe that this is closer to the original source code, as the
previous code was rife with duplication and redundant conditions.

I've also modified the alignment so that it works correctly on
platforms where the alignment requirement is larger than 4. This is
achieved using a C++11 definition. Without this correction, the
address sanitizer on x86-64 Linux throws a fit.

I've also renamed the 'clean storage' functions to make their
differences more clear, as 'ClearEmptyStorage' and
'ClearUnusedStorage' don't get the point across very well. This can
be reverted if those were actually the original names of the
functions in the original source code.

By making this modification, an invalid memory access that occurs
when a storage becomes full and must be cleaned is avoided. This
memory access would be also be detected by the address-sanitizer.