Make memory allocator code more accurate.
This code produces identical assembly code to that found in the
latest Steam EXE (with one exception).

You can easily test the accuracy of code yourself at godbolt.org,
with the compiler set to 'MSVC 19.0 x86' and the compiler options set
to `/O2 /Gr /Oy- /MD`.

The only code that isn't absolutely identical to the original is the
two instances of the line `processedStorage += size;` in
`DefragmentAndGarbageCollectStorage`: in the original EXE, both
instances produce their own assembly code, while, with this
decompiled version, the compiler merges them together as an
optimisation by exploiting code flow.

Still, despite this minor difference in optimisation, this code
should behaviour indentically to the original, fixing any bugs
and other mistakes introduced during the decompilation process.