Set SDL2 render clear colour.
`SDL_RenderClear` uses the colour specified by
`SDL_SetRenderDrawColor`, which is incidentally set by the screen
dimming code. Framebuffers usually ignore the alpha component so this
is a very minor fix, but I think it's worth explicitly setting the
colour since that what's meant to be done, and it avoids side-effects
that would occur if a modder or whoever changed the dimming to a
colour besides black.