Fix potential allocator corruption.
If `AllocateStorage` is used to allocate to a local variable, then it
MUST be manually freed with `RemoveStorageEntry`!

If this isn't done, then dangling pointers to the stack are left in
the storage's allocation list. Not only does this lead to
'use-after-scope' errors in Linux's address sanitizer, but this can
also lead to corruption:

The reason for this is that it's possible for the data in the stack
that is pointed at by one of these dangling pointers to be filled
with the address of another allocation. If this other allocation is
freed with `RemoveStorageEntry`, then it will free the wrong
allocation.

There are other instances of local allocations not being freed
elsewhere in this codebase, and these should really be fixed...