Revert "UserAchievements: Free local variables"
Turns out allocated Strings in v5/Mania NEVER get free'd manually, and rely on
the garbage collector.

Most STR allocations are done on local variables stored on the stack.
Trying to free any String will iterate over the allocated STR entries and try
dereferencing a pointer to a stack variable that is long gone, which will
potentially corrupt data if that pointer ends up being filled with the address
of another allocation.

Back to the reverted commit: not only this 'fix' would trigger the bug, but it
would also try to free the string rather than the underlying `chars` buffer,
which would match absolutely no entry.

Let's stay faithful to the original and don't free the strings, and hope that
the STR memory pool never needs to be garbage collected...

This reverts commit 6e13c8059868639d6c1886ff29bd5a6813b6a5c7.